More simplifications
