--- internal/domain/errors.go
@@ -1,3 +1,0 @@
-package domain
-
-var ()
